Finish game implementation in UI
Game now runs completely fine fully using the custom made buttons in the
UI instead of the prompt by js. Issue to be solved later is how to
improve the overall flow of the code.
So, what actually is being used now for the playRound is a bit weird,
you take a choice from the computer, take the human, and run the
comparison from the eventListener of the choiceButtons in the
humanChoice function. This approach is a bit unreasonable, in my
opinion.